+++ lib/legion/extensions/smtp/runners/email.rb
@@ -8,7 +8,7 @@ module Legion
       module Runners
         module Email
           def send(to:, body:, from:, subject:, port: 25, address: 'localhost', **_opts) # rubocop:disable Metrics/ParameterLists
-            Net::SMTP.start(address, port) do |smtp|
+            connection(address: address, port: port).start do |smtp|
               smtp.send_message body, from, to, subject
             end
           end